# account-service/api/account.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/account.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Account_Service"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/account_service/CreateBuisness": {
      "post": {
        "operationId": "Account_Service_CreateBuisness",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateBuisnessResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateBuisnessRequest"
            }
          }
        ],
        "tags": [
          "Account_Service"
        ]
      }
    },
    "/api/account_service/GetUserProfile/{uuid}": {
      "get": {
        "operationId": "Account_Service_GetUserProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetUserProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Account_Service"
        ]
      }
    },
    "/api/account_service/UpdateUser/{uuid}": {
      "post": {
        "operationId": "Account_Service_UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiUpdateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Account_ServiceUpdateUserBody"
            }
          }
        ],
        "tags": [
          "Account_Service"
        ]
      }
    }
  },
  "definitions": {
    "Account_ServiceUpdateUserBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "apiCreateBuisnessRequest": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "apiCreateBuisnessResponse": {
      "type": "object"
    },
    "apiCreateUserResponse": {
      "type": "object"
    },
    "apiDeleteUserResponse": {
      "type": "object"
    },
    "apiGetBuisnessResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "apiGetUserProfileResponse": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "apiPingResponse": {
      "type": "object",
      "properties": {
        "pingMessage": {
          "type": "string"
        }
      }
    },
    "apiUpdateUserResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}